requirements.txt
asgiref==3.11.0
certifi==2026.2.25
charset-normalizer==3.4.6
Django==5.2.11
django-environ==0.13.0
idna==3.11
psycopg==3.2.3
requests==2.32.5
sqlparse==0.5.5
typing_extensions==4.15.0
urllib3==2.6.3
